requirements.txt
python-dotenv==0.20.0
redis==4.2.2
paramiko==2.11.0
psycopg2==2.9.3
celery==5.2.7
prometheus-api-client==0.5.1
pandas==1.4.3
rq==1.11.0
pipe==2.0
